pyproject.toml
[tool.poetry]
name = "deploy-functions-oidc-template"
version = "1.6.0"
description = "Cognite Functions template for OIDC projects"
authors = ["hakon.treider@cognite.com"]

[tool.poetry.dependencies]
python = "^3.9"
pytest-custom-exit-code = ">=0.3"
cognite-sdk = ">=6"
prophet = "^1.1.4"
statsmodels = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = ">=7"
pre-commit = ">=3"

[tool.isort]
profile = "black"
force_single_line = false
atomic = true
lines_after_imports = 2
lines_between_types = 1
line_length = 120
known_third_party = ["cognite","pytest"]

[tool.black]
line-length = 120
